Make slack loops slack coils.
With good tracking of the cable path, aerial slack loops that use snowshoes will be naturally represented by the path of the cable and tracking of footage to a physical location will work as it should. No need to overcomplicate aerial figure 8/snowshoe slack sections by trying to capture them as a special object.